Created a MonitorGrid model for holding all of the specific logic around it; removed the WorkspaceGrid entirely since it turned out to be unnecessary.

+++ src/models/monitor.rs
@@ -2,13 +2,13 @@ use super::Window;
 
 #[derive(Clone, Debug)]
 pub struct Monitor {
-    pub width: usize,
-    pub height: usize,
+    pub width: i32,
+    pub height: i32,
     pub windows: Vec<Window>,
 }
 
 impl Monitor {
-    pub const fn new(width: usize, height: usize) -> Self {
+    pub const fn new(width: i32, height: i32) -> Self {
         Self {
             width,
             height,
